Create index.html
The home page for my ASE coursework where I had to make a fake uber website with Google Maps API, a front end and a back end linking to MongoDB and It was my first time practically using Mocha as part of a TDD process.  The app is quite simple as the cars do not follow road paths but it was a fun challenge to make a full stack application.